#VYE-325 #comment combine review and confirmation pages into one (#28981)
