+++ src/applications/verify-your-enrollment/containers/VerificationReviewWrapper.jsx
@@ -4,7 +4,6 @@ import { useHistory } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import {
-  VaButtonPair,
   VaRadio,
   VaRadioOption,
 } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
@@ -13,50 +12,67 @@ import { useScrollToTop } from '../hooks/useScrollToTop';
 import VerifyEnrollmentStatement from '../components/VerifyEnrollmentStatement';
 import EnrollmentCard from '../components/Enrollmentcard';
 import NeedHelp from '../components/NeedHelp';
-import {
-  VERIFICATION_RELATIVE_URL,
-  CONFIRMATION_REVIEW_RELATIVE_URL,
-} from '../constants';
-import { updateToggleEnrollmentCard } from '../actions';
+import { VERIFICATION_RELATIVE_URL } from '../constants';
 import Loader from '../components/Loader';
 import { useData } from '../hooks/useData';
+import {
+  updateToggleEnrollmentSuccess,
+  updatePendingVerifications,
+  updateVerifications,
+  verifyEnrollmentAction,
+} from '../actions';
 
 const VerificationReviewWrapper = ({
   children,
   enrollmentData,
-  dispatchupdateToggleEnrollmentCard,
+  // loggedIEnenrollmentData,
+  dispatchUpdateToggleEnrollmentSuccess,
+  dispatchUpdatePendingVerifications,
+  dispatchUpdateVerifications,
+  dispatchVerifyEnrollmentAction,
+  // isUserLoggedIn,
+  // dispatchupdateToggleEnrollmentCard,
 }) => {
   useScrollToTop();
-  const [radioValue, setRadioValue] = useState(null);
+  const [radioValue, setRadioValue] = useState(false);
   const [errorStatement, setErrorStatement] = useState(null);
   const { loading } = useData();
   const [enrollmentPeriodsToVerify, setEnrollmentPeriodsToVerify] = useState(
     [],
   );
+  const [currentPendingAwardIDs, setCurrentPendingAwardIDs] = useState([]);
+  // const userData = isUserLoggedIn ? loggedIEnenrollmentData : enrollmentData;
   const history = useHistory();
 
   const handleBackClick = () => {
     history.push(VERIFICATION_RELATIVE_URL);
   };
-
-  const handleNextClick = () => {
-    if (radioValue === null) {
-      setErrorStatement('Please provide a response');
-    } else {
-      history.push(CONFIRMATION_REVIEW_RELATIVE_URL);
-    }
-  };
-
   const handleRadioClick = e => {
     const { value } = e.detail;
     setRadioValue(value);
     setErrorStatement(null);
-    if (value === 'true') {
-      dispatchupdateToggleEnrollmentCard(true);
-    }
-    if (value === 'false') {
-      dispatchupdateToggleEnrollmentCard(false);
-    }
+  };
+
+  // used with mock data to mock what happens after
+  // successfully verifying
+  const handleVerification = () => {
+    const currentDateTime = new Date().toISOString();
+    // update awardIds to a blank array
+    dispatchUpdatePendingVerifications({ awardIds: [] });
+    const newVerifiedIDS = currentPendingAwardIDs?.map(id => {
+      return {
+        PendingVerificationSubmitted: currentDateTime,
+        awardIds: [id],
+      };
+    });
+    dispatchUpdateVerifications(newVerifiedIDS);
+    dispatchVerifyEnrollmentAction();
+  };
+
+  const handleSubmission = () => {
+    handleVerification();
+    dispatchUpdateToggleEnrollmentSuccess(true);
+    history.push(VERIFICATION_RELATIVE_URL);
   };
 
   useEffect(
@@ -71,6 +87,7 @@ const VerificationReviewWrapper = ({
         // add all previouslyVerified data into single array
         const { awardIds } = pendingVerifications;
         const toBeVerifiedEnrollmentsArray = [];
+        setCurrentPendingAwardIDs(awardIds);
         awardIds.forEach(id => {
           // check for each id inside award_ids array
           if (awards.some(award => award.id === id)) {
@@ -145,25 +162,33 @@ const VerificationReviewWrapper = ({
                       tile
                       value="true"
                     />
-                    <VaRadioOption
-                      id="vye-radio-button-no"
-                      label="No, this information isn't correct."
-                      name="vye-radio-group1"
-                      tile
-                      value="false"
-                    />
                   </VaRadio>
                 </div>
                 <div
                   style={{
                     paddingLeft: '8px',
+                    marginTop: '24px',
+                    display: 'flex',
+                    columnGap: '10px',
                   }}
                 >
-                  <VaButtonPair
-                    continue
-                    onPrimaryClick={handleNextClick}
-                    onSecondaryClick={handleBackClick}
-                  />
+                  <va-button onClick={handleBackClick} back uswds />
+                  {radioValue && (
+                    <va-button
+                      onClick={handleSubmission}
+                      text="Submit"
+                      submit
+                      uswds
+                    />
+                  )}
+                  {!radioValue && (
+                    <va-button
+                      onClick={handleSubmission}
+                      text="Submit"
+                      disabled
+                      uswds
+                    />
+                  )}
                 </div>
               </>
             )}
@@ -182,14 +207,22 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = {
-  dispatchupdateToggleEnrollmentCard: updateToggleEnrollmentCard,
+  dispatchUpdatePendingVerifications: updatePendingVerifications,
+  dispatchUpdateToggleEnrollmentSuccess: updateToggleEnrollmentSuccess,
+  dispatchUpdateVerifications: updateVerifications,
+  dispatchVerifyEnrollmentAction: verifyEnrollmentAction,
 };
 
 VerificationReviewWrapper.propTypes = {
   children: PropTypes.any,
-  dispatchupdateToggleEnrollmentCard: PropTypes.func,
+  dispatchUpdatePendingVerifications: PropTypes.func,
+  dispatchUpdateToggleEnrollmentSuccess: PropTypes.func,
+  dispatchUpdateVerifications: PropTypes.func,
+  dispatchVerifyEnrollmentAction: PropTypes.func,
   enrollmentData: PropTypes.object,
+  isUserLoggedIn: PropTypes.bool,
   link: PropTypes.func,
+  loggedIEnenrollmentData: PropTypes.object,
 };
 export default connect(
   mapStateToProps,
