+++ src/applications/verify-your-enrollment/routes.jsx
@@ -3,13 +3,11 @@ import { Route, Switch } from 'react-router-dom';
 import {
   BENEFITS_PROFILE_RELATIVE_URL,
   VERIFICATION_REVIEW_RELATIVE_URL,
-  CONFIRMATION_REVIEW_RELATIVE_URL,
 } from './constants';
 // import App from './containers/App';
 import EnrollmentVerificationPageWrapper from './containers/EnrollmentVerificationPageWrapper';
 import BenefitsProfileWrapper from './containers/BenefitsProfilePageWrapper';
 import VerificationReviewWrapper from './containers/VerificationReviewWrapper';
-import ConfirmationReviewWrapper from './containers/ConfirmationReviewWrapper';
 
 const routes = (
   <Switch>
@@ -29,13 +27,6 @@ const routes = (
       path={`${VERIFICATION_REVIEW_RELATIVE_URL}`}
     >
       <VerificationReviewWrapper />
-    </Route>
-    <Route
-      exact
-      key="VerificationReview"
-      path={`${CONFIRMATION_REVIEW_RELATIVE_URL}`}
-    >
-      <ConfirmationReviewWrapper />
     </Route>
   </Switch>
 );
